projects/elevenlabs-voice-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elevenlabs==0.2.26
openai==1.3.0
pyaudio==0.1.23
soundfile==0.12.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
pydub==0.25.1
librosa==0.10.1
scipy==1.11.3
python-multipart==0.0.6
jinja2==3.1.2
